# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Minions Market API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}/products:
    get:
      operationId: getUserProducts
      tags: [users]
      summary: Get user products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsListResponse"
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, cheapest, expensive, popular]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /products/{id}/favorite:
    post:
      operationId: toggleFavorite
      tags: [products]
      summary: Toggle favorite
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  favorited:
                    type: boolean
                required: [favorited]
...
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: List all products (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsListResponse"
...
  /admin/products/{id}:
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      summary: Delete product permanently (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        avatar:
          type: string
        bio:
          type: string
        balance:
          type: number
        frozenBalance:
          type: number
        totalSales:
          type: integer
        totalPurchases:
          type: integer
        totalVolume:
          type: number
        rating:
          type: number
        reviewCount:
          type: integer
        isAdmin:
          type: boolean
        isVerified:
          type: boolean
        isBanned:
          type: boolean
        sellerLevel:
          type: string
        telegramId:
          type: string
        refCode:
          type: string
        createdAt:
          type: integer
        lastActive:
          type: integer
      required: [id, username]
...
    Product:
      type: object
      properties:
        id:
          type: string
        sellerId:
          type: string
        seller:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        categoryName:
          type: string
        subcategory:
          type: string
        images:
          type: array
          items:
            type: string
        deliveryType:
          type: string
          enum: [auto, manual]
        game:
          type: string
        server:
          type: string
        status:
          type: string
        views:
          type: integer
        soldCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        isFavorited:
          type: boolean
        isPromoted:
          type: boolean
        createdAt:
          type: integer
        updatedAt:
          type: integer
      required: [id, sellerId, title, price, category]
    CreateProductBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        subcategory:
          type: string
        images:
          type: array
          items:
            type: string
        deliveryType:
          type: string
          enum: [auto, manual]
        deliveryData:
          type: string
        game:
          type: string
        server:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [title, description, price, category]
    UpdateProductBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        subcategory:
          type: string
        images:
          type: array
          items:
            type: string
        deliveryType:
          type: string
          enum: [auto, manual]
        deliveryData:
          type: string
        game:
          type: string
        server:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
    ProductsListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [products, total]
...